% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msand-package.R
\docType{package}
\name{msand-package}
\alias{msand}
\alias{msand-package}
\title{msand: Personal Package for Visualizations and RMarkdown Templates}
\description{
Custom colour palette and themes for visualization. Custom rmarkdown templates for presentations, github_documentation html_documentation, PDFs, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MichaelSandilands/msand}
  \item Report bugs at \url{https://github.com/MichaelSandilands/msand/issues}
}

}
\keyword{internal}
