% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scale_manual.R
\name{palette_msand}
\alias{palette_msand}
\title{msand palettes for use with scales}
\usage{
palette_msand()
}
\description{
These palettes are mainly called internally by msand \verb{scale_*_msand()} functions.
}
\examples{
library(scales)
scales::show_col(palette_msand())

}
