% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme_msand.R
\name{theme_msand}
\alias{theme_msand}
\title{msand themes for ggplot2.}
\usage{
theme_msand(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
The \code{theme_msand()} function creates a custom theme using msand colors.
}
\examples{
library(ggplot2)

diamonds \%>\%
    ggplot(aes(carat, price, colour = cut)) +
    geom_point() +
    scale_colour_msand() +
    theme_msand()

}
\seealso{
\code{\link[=scale_manual]{scale_manual()}}
}
